+++ mobile/android/android-components/components/browser/engine-gecko-nightly/src/test/java/mozilla/components/browser/engine/gecko/GeckoEngineTest.kt
@@ -105,7 +105,8 @@ class GeckoEngineTest {
                 TrackingProtectionPolicy.SOCIAL,
                 TrackingProtectionPolicy.ANALYTICS,
                 TrackingProtectionPolicy.CONTENT,
-                TrackingProtectionPolicy.TEST
+                TrackingProtectionPolicy.TEST,
+                TrackingProtectionPolicy.CRYPTOMINING
         ).categories, contentBlockingSettings.categories)
         assertEquals(defaultSettings.trackingProtectionPolicy, TrackingProtectionPolicy.all())
 
@@ -148,7 +149,8 @@ class GeckoEngineTest {
             TrackingProtectionPolicy.SOCIAL,
             TrackingProtectionPolicy.ANALYTICS,
             TrackingProtectionPolicy.CONTENT,
-            TrackingProtectionPolicy.TEST
+            TrackingProtectionPolicy.TEST,
+            TrackingProtectionPolicy.CRYPTOMINING
         ).categories, contentBlockingSettings.categories)
         assertTrue(engine.settings.testingModeEnabled)
         assertEquals("test-ua", engine.settings.userAgentString)