+++ mobile/android/android-components/docs/changelog.md
@@ -37,6 +37,9 @@ permalink: /changelog/
   * Added option to add "edit actions" that will show up next to the URL in edit mode.
   * Added option to set a listener for clicks on the site security indicator (globe / lock icon).
 
+* **browser-engine-gecko-nightly**
+  * Added new `TrackingProtectionPolicy` category for blocking cryptocurrency miners (`TrackingProtectionPolicy.CRYPTOMINING`).
+
 # 0.48.0
 
 * [Commits](https://github.com/mozilla-mobile/android-components/compare/v0.47.0...v0.48.0)